frontend, routing and api fixes, plus new angular features
changed the way the frontend looks, and added some new functionality to it. Changed the way it is styled, now using FlexLayout for positioning the devices. Added angular materials to the frontend environment, to allow for a better UI experience. updated a lot of packages. next up is integrating with HomeBridge and modularizing the frontend with device-specific controls